refactor: extract host rsvp endpoint into _event_host_rsvps.py
