+++ backend/tests/test_host_rsvp.py
@@ -228,7 +228,7 @@ class TestSetGuestRsvp:
     def test_broadcasts_capacity_change_excluding_actor(
         self, api_client, host_rsvp_event, host_user, guest
     ):
-        with patch("community._event_host_actions.broadcast_capacity_change") as mock_broadcast:
+        with patch("community._event_host_rsvps.broadcast_capacity_change") as mock_broadcast:
             api_client.post(
                 f"/api/community/events/{host_rsvp_event.id}/rsvps/{guest.pk}/rsvp/",
                 {"status": RSVPStatus.ATTENDING},
